build fix for 5452fae33c0baeff82dfb056dee433bab2ef687d

+++ services/graphics-server/src/backend/betrusted.rs
@@ -5,8 +5,8 @@ use utralib::generated::*;
 use xous::MemoryRange;
 
 pub const FB_WIDTH_WORDS: usize = 11;
-pub const FB_WIDTH_PIXELS: usize = WIDTH;
-pub const FB_LINES: usize = LINES;
+pub const FB_WIDTH_PIXELS: usize = WIDTH as usize;
+pub const FB_LINES: usize = LINES as usize;
 pub const FB_SIZE: usize = FB_WIDTH_WORDS * FB_LINES; // 44 bytes by 536 lines
 const CONFIG_CLOCK_FREQUENCY: u32 = 100_000_000;
